+++ src/app/persons/experiences/experiences.component.html
@@ -1,6 +1,6 @@
 <nz-layout>
     <nz-content>
         <h1>Work experience of {{editedPerson?.name}} {{editedPerson?.surname}}</h1>
-        <app-experience-form [person]="editedPerson"></app-experience-form>
+        <app-experience-form [person]="editedPerson" (submitted)="submitExperience($event)"></app-experience-form>
     </nz-content>
 </nz-layout>
